CN-4578: refactoring contacts logic
