+++ app/redux/actions.js
@@ -22,4 +22,5 @@ export * from "./celPay/celPayActions";
 export * from "./graph/graphActions";
 export * from "./community/communityActions";
 export * from "./compliance/complianceActions";
+export * from "./contacts/contactsActions";
 // NOTE(fj): plop reduxGen importing new Actions here
