+++ app/redux/contacts/contactsReducer.js
@@ -1,0 +1,18 @@
+import ACTIONS from "../../constants/ACTIONS";
+
+const initialState = {
+  contacts: [],
+};
+
+export default function contactsReducer(state = initialState, action) {
+  switch (action.type) {
+    case ACTIONS.GET_CONNECTED_CONTACTS_SUCCESS:
+      return {
+        ...state,
+        contacts: [...action.contacts.friendsWithApp],
+      };
+
+    default:
+      return { ...state };
+  }
+}
